feat(browser): fixed code created 2 files for file_storage
